move GA to layout. component folders
