+++ web/src/app/page.tsx
@@ -1,13 +1,13 @@
+"use client";
+
 import React from "react";
-import WikipediaBrowser from "../components/WikipediaBrowser";
-import { GoogleAnalytics } from "@next/third-parties/google";
+import WikipediaBrowser from "../components/WikipediaBrowser/WikipediaBrowser";
 
 
 const Home: React.FC = () => {
   return (
     <div>
       <WikipediaBrowser />
-      <GoogleAnalytics gaId="G-XFQJZD9P2K" />
     </div>
   );
 };
